Revert "chore: migrate `tools` to TypeScript (#296)"
This reverts commit b03f7413cc4740b05333ef49dfcf3e9a12fd5c44.
