+++ .gitignore
@@ -7,4 +7,3 @@ package-lock.json
 build/
 .eslintcache
 !packages/cli/src/commands/init/__fixtures__/editTemplate/node_modules
-*.tsbuildinfo
