--- .vscode/extensions.json
@@ -1,10 +1,0 @@
-{
-  // See http://go.microsoft.com/fwlink/?LinkId=827846
-  // for the documentation about the extensions.json format
-  "recommendations": [
-    "dbaeumer.vscode-eslint",
-    "redhat.vscode-yaml",
-    "flowtype.flow-for-vscode",
-    "esbenp.prettier-vscode"
-  ]
-}
