+++ jest.config.js
@@ -12,7 +12,7 @@ module.exports = {
       ...common,
       displayName: 'unit',
       setupFiles: ['<rootDir>/jest/setupUnitTests.js'],
-      testMatch: ['<rootDir>/**/__tests__/*{.,-}test.[jt]s'],
+      testMatch: ['<rootDir>/**/__tests__/*{.,-}test.js'],
     },
   ],
 };
