+++ package.json
@@ -4,16 +4,14 @@
     "packages/*"
   ],
   "scripts": {
-    "prebuild": "yarn build:ts",
     "build": "node ./scripts/build.js",
-    "build:ts": "node ./scripts/buildTs.js",
-    "build-clean": "rm -rf ./packages/*/build ./packages/*/tsconfig.tsbuildinfo",
+    "build-clean": "rm -rf ./packages/*/build",
     "watch": "node ./scripts/watch.js",
     "test": "jest",
     "test:ci:unit": "jest packages --ci --coverage",
     "test:ci:e2e": "jest e2e --ci -i",
-    "lint": "eslint --ext .js,.ts . --cache --report-unused-disable-directives",
     "test:ci:cocoapods": "ruby packages/platform-ios/native_modules.rb",
+    "lint": "eslint . --cache --report-unused-disable-directives",
     "flow-check": "flow check",
     "postinstall": "yarn build",
     "publish": "yarn build-clean && yarn build && lerna publish"
@@ -24,16 +22,12 @@
     "@babel/plugin-transform-strict-mode": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
     "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-typescript": "^7.3.3",
-    "@react-native-community/eslint-config": "^0.0.5",
-    "@types/jest": "^24.0.11",
-    "@types/node": "^11.13.0",
+    "@react-native-community/eslint-config": "^0.0.3",
     "babel-jest": "^24.6.0",
     "babel-plugin-module-resolver": "^3.2.0",
     "chalk": "^2.4.2",
     "eslint": "^5.10.0",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-import-resolver-typescript": "^1.1.1",
     "eslint-plugin-import": "^2.17.0",
     "execa": "^1.0.0",
     "flow-bin": "^0.97.0",
@@ -44,7 +38,6 @@
     "metro-memory-fs": "^0.53.1",
     "micromatch": "^3.1.10",
     "mkdirp": "^0.5.1",
-    "string-length": "^2.0.0",
-    "typescript": "^3.4.5"
+    "string-length": "^2.0.0"
   }
 }
