+++ packages/cli/package.json
@@ -49,6 +49,7 @@
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
     "morgan": "^1.9.0",
+    "node-fetch": "^2.2.0",
     "node-notifier": "^5.2.1",
     "open": "^6.2.0",
     "ora": "^3.4.0",
