+++ packages/platform-android/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@react-native-community/cli-tools": "^2.0.0-alpha.20",
     "logkitty": "^0.4.0",
+    "node-fetch": "^2.2.0",
     "slash": "^2.0.0",
     "xmldoc": "^0.4.0"
   },
