+++ packages/tools/package.json
@@ -5,13 +5,7 @@
   "dependencies": {
     "chalk": "^1.1.1",
     "lodash": "^4.17.5",
-    "mime": "^2.4.1",
-    "node-fetch": "^2.5.0"
-  },
-  "devDependencies": {
-    "@types/lodash": "^4.14.123",
-    "@types/mime": "^2.0.1",
-    "@types/node-fetch": "^2.3.3"
+    "mime": "^1.3.4"
   },
   "files": [
     "build"
