+++ packages/tools/src/__mocks__/logger.js
@@ -4,6 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
+ * @format
  */
 
 module.exports.out = () => jest.fn();
