+++ packages/tools/src/__tests__/groupFilesByType-test.js
@@ -4,6 +4,8 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
+ * @format
+ * @emails oncall+javascript_foundation
  */
 
 import groupFilesByType from '../groupFilesByType';
