+++ packages/tools/src/getDefaultUserTerminal.js
@@ -1,4 +1,4 @@
-const getDefaultUserTerminal = (): string | undefined =>
+const getDefaultUserTerminal = (): ?string =>
   process.env.REACT_TERMINAL || process.env.TERM_PROGRAM;
 
 export default getDefaultUserTerminal;
