+++ packages/tools/src/groupFilesByType.js
@@ -4,10 +4,21 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
+ * @flow
  */
 
 import {groupBy} from 'lodash';
 import mime from 'mime';
+
+/**
+ * Since there are no officially registered MIME types
+ * for ttf/otf yet http://www.iana.org/assignments/media-types/media-types.xhtml,
+ * we define two non-standard ones for the sake of parsing
+ */
+mime.define({
+  'font/opentype': ['otf'],
+  'font/truetype': ['ttf'],
+});
 
 /**
  * Given an array of files, it groups it by it's type.
@@ -21,5 +32,5 @@ import mime from 'mime';
  * the returned object will be: {font: ['fonts/a.ttf'], image: ['images/b.jpg']}
  */
 export default function groupFilesByType(assets: Array<string>) {
-  return groupBy(assets, type => (mime.getType(type) || '').split('/')[0]);
+  return groupBy(assets, type => mime.lookup(type).split('/')[0]);
 }
