+++ packages/tools/src/index.js
@@ -1,3 +1,6 @@
+/**
+ * @flow
+ */
 export {default as logger} from './logger';
 export {default as groupFilesByType} from './groupFilesByType';
 export {default as isPackagerRunning} from './isPackagerRunning';
