+++ packages/tools/src/logger.js
@@ -1,3 +1,6 @@
+/**
+ * @flow
+ */
 import chalk from 'chalk';
 
 const SEPARATOR = ', ';
