+++ scripts/build.js
@@ -25,20 +25,30 @@ const mkdirp = require('mkdirp');
 const babel = require('@babel/core');
 const chalk = require('chalk');
 const micromatch = require('micromatch');
-const {
-  PACKAGES_DIR,
-  getPackages,
-  OK,
-  adjustToTerminalWidth,
-} = require('./helpers');
+const stringLength = require('string-length');
+const {PACKAGES_DIR, getPackages} = require('./helpers');
 
+const OK = chalk.reset.inverse.bold.green(' DONE ');
 const SRC_DIR = 'src';
 const BUILD_DIR = 'build';
 const JS_FILES_PATTERN = '**/*.js';
-const TS_FILE_PATTERN = '**/*.ts';
 const IGNORE_PATTERN = '**/__{tests,mocks,fixtures}__/**';
 
 const transformOptions = require('../babel.config.js');
+
+const adjustToTerminalWidth = str => {
+  const columns = process.stdout.columns || 80;
+  const WIDTH = columns - stringLength(OK) + 1;
+  const strs = str.match(new RegExp(`(.{1,${WIDTH}})`, 'g'));
+  let lastString = strs[strs.length - 1];
+  if (lastString.length < WIDTH) {
+    lastString += Array(WIDTH - lastString.length).join(chalk.dim('.'));
+  }
+  return strs
+    .slice(0, -1)
+    .concat(lastString)
+    .join('\n');
+};
 
 function getPackageName(file) {
   return path.relative(PACKAGES_DIR, file).split(path.sep)[0];
@@ -49,7 +59,7 @@ function getBuildPath(file, buildFolder) {
   const pkgSrcPath = path.resolve(PACKAGES_DIR, pkgName, SRC_DIR);
   const pkgBuildPath = path.resolve(PACKAGES_DIR, pkgName, buildFolder);
   const relativeToSrcPath = path.relative(pkgSrcPath, file);
-  return path.resolve(pkgBuildPath, relativeToSrcPath).replace(/\.ts$/, '.js');
+  return path.resolve(pkgBuildPath, relativeToSrcPath);
 }
 
 function buildNodePackage(p) {
@@ -79,10 +89,7 @@ function buildFile(file, silent) {
 
   mkdirp.sync(path.dirname(destPath), '777');
 
-  if (
-    !micromatch.isMatch(file, JS_FILES_PATTERN) &&
-    !micromatch.isMatch(file, TS_FILE_PATTERN)
-  ) {
+  if (!micromatch.isMatch(file, JS_FILES_PATTERN)) {
     fs.createReadStream(file).pipe(fs.createWriteStream(destPath));
     silent ||
       process.stdout.write(
